+++ debug_harness/Parser/SemanticCheck/SymTab/SymTab.h
@@ -44,7 +44,7 @@ int PushVarOntoScope(SymTab_t *symtab, enum VarType var_type, char *id);
 /* Pushes a new array onto the current scope (head) */
 int PushArrayOntoScope(SymTab_t *symtab, enum VarType var_type, char *id, int start, int end, int element_size);
 
-int PushConstOntoScope(SymTab_t *symtab, char *id, int value);
+int PushConstOntoScope(SymTab_t *symtab, char *id, long long value);
 
 /* Pushes a new procedure onto the current scope (head) */
 /* NOTE: args can be NULL to represent no args */
